review fixes. Mostly moving stuff to other files.
