+++ lib/sycamore/sycamore/tests/unit/transforms/test_detr_partitioner.py
@@ -1,7 +1,7 @@
 from sycamore.data import Element
 from sycamore.transforms.detr_partitioner import SycamorePDFPartitioner, DeformableDetr
 from sycamore.data import BoundingBox
-from sycamore.tests.unit.transforms.compare_detr_impls import compare_batched_sequenced, deep_eq
+from sycamore.tests.unit.transforms.compare_detr_impls import compare_batched_sequenced
 
 from PIL import Image
 import json
@@ -51,18 +51,6 @@ class TestSycamorePDFPartitioner:
             for result in results:
                 for element in result:
                     json.dumps(element.properties)
-
-    def test_deep_eq(self):
-        class Tmp:
-            def __init__(self, x, y):
-                self.x = x
-                self.y = y
-
-            def doit():
-                pass
-
-        assert deep_eq(Tmp(1, 1), Tmp(1, 1))
-        assert not deep_eq(Tmp(1, 1), Tmp(1, 2))
 
     def test_batched_sequenced(self):
         s = SycamorePDFPartitioner("Aryn/deformable-detr-DocLayNet")
